+++ src/test/java/com/pexapark/ElectricityFarm/CapacityServiceServiceITest.java
@@ -4,6 +4,7 @@ import com.github.database.rider.core.api.configuration.DBUnit;
 import com.github.database.rider.core.api.dataset.ExpectedDataSet;
 import com.github.database.rider.spring.api.DBRider;
 import com.pexapark.ElectricityFarm.config.IntegrationTestDB;
+import com.pexapark.ElectricityFarm.config.test.CleanDatabaseBeforeAndAfter;
 import com.pexapark.ElectricityFarm.model.TimeRange;
 import com.pexapark.ElectricityFarm.services.CapacityFactorService;
 import org.assertj.core.api.Assertions;
@@ -11,6 +12,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.annotation.IfProfileValue;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @IfProfileValue(name = "spring.profiles.active", value = "integration")
@@ -26,6 +28,7 @@ public class CapacityServiceServiceITest extends IntegrationTestDB {
    *        Example: @CleanDatabaseBeforeAndAfter(value = "dbrider_data/expextedLiquibase_intialCapacityService.json")
    **/
   @Test
+//  @CleanDatabaseBeforeAndAfter(value = "dbrider_data/expextedLiquibase_intialCapacityService.json")
   @ExpectedDataSet(value = "dbrider_data/expextedLiquibase_intialCapacityService.json")
   public void test_testing() {
     TimeRange timeRange =
@@ -37,6 +40,6 @@ public class CapacityServiceServiceITest extends IntegrationTestDB {
 
     Assertions.assertThat(
             capacityFactorService.calcCapacityFactorByTimeRange(electricityFarmId, timeRange))
-        .isEqualTo(6.17);
+        .isEqualTo(BigDecimal.valueOf(6.17));
   }
 }
